scx_wd40: factor out task placement
