+++ scheds/rust/scx_wd40/build.rs
@@ -11,6 +11,7 @@ fn main() {
         .add_source("src/bpf/lb_domain.bpf.c")
         .add_source("src/bpf/common.bpf.c")
         .add_source("src/bpf/deadline.bpf.c")
+        .add_source("src/bpf/placement.bpf.c")
         .add_source("../../../lib/sdt_task.bpf.c")
         .add_source("../../../lib/sdt_alloc.bpf.c")
         .compile_link_gen()
