+++ scheds/rust/scx_wd40/src/bpf/common.bpf.c
@@ -21,6 +21,8 @@
 const volatile u32 dom_numa_id_map[MAX_DOMS];
 const volatile u32 debug;
 const volatile u32 load_half_life = 1000000000	/* 1s */;
+const volatile u32 nr_doms = 32;	/* !0 for veristat, set during init */
+const volatile u32 nr_nodes = 32;	/* !0 for veristat, set during init */
 
 /* base slice duration */
 volatile u64 slice_ns;
